change structure of cli functions; move functions from core_functions.py to cli.py
