+++ docs/source/developer.rst
@@ -1,0 +1,3 @@
+*****************************
+Documentation for developers
+*****************************
