+++ docs/source/installation.rst
@@ -1,5 +1,5 @@
-------------
+**********************
 Installation
-------------
+**********************
 
 To install BABS: ___________
