+++ docs/source/overview.rst
@@ -1,0 +1,3 @@
+**********************
+Overview of BABS
+**********************
