+++ docs/source/preparation.rst
@@ -1,0 +1,9 @@
+**********************
+Get prepared
+**********************
+
+We'll cover these three aspects:
+
+- Input dataset
+- Container (datalad dataset)
+- Config yaml file
