+++ docs/source/walkthrough.rst
@@ -1,0 +1,3 @@
+**********************
+Example walkthrough
+**********************
