+++ setup.cfg
@@ -47,9 +47,9 @@ include_package_data = True
 
 [options.entry_points]
 console_scripts =
-    babs-init=babs.cli:babs_init_cli
-    babs-submit=babs.cli:babs_submit_cli
-    babs-status=babs.cli:babs_status_cli
+    babs-init=babs.cli:babs_init_main
+    babs-submit=babs.cli:babs_submit_main
+    babs-status=babs.cli:babs_status_main
 
 [flake8]
 ignore = E226,E302,E41,E731,E123,W503
